reorg modules, add comments
